+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 
-## [2.0.1]
+## [2.0.2]
 
 ### - 2024-09-24
 - annona/file/Upload.class 파일명 암호화 Deprecated 관련 패치
+- annona/file/*.class 자잘한 코딩 패치 작업
 
 ### - 2024-09-13
 - HttpRequest class v1.1.0 update